#463: SpielerOffsetDialog renamed to PlayerSubskillOffsetDialog allows adjusting experience subskill
