+++ src/main/resources/release_notes.md
@@ -30,9 +30,12 @@ Changelist HO! 3.0
 
 - [NEW] Full control on which game to download (e.g. exclude HTO integrated games)  #290
 
+
 ### Squad
 
   - [FIX] ordering of best position is now saved on closing #397
+  - [NEW] manual adjustment of experience level (similar to other skills) #463
+
 
 ### Team Analyser
 
@@ -44,7 +47,6 @@ Changelist HO! 3.0
  - [NEW] impact of match orders on rating is now directly visible from the subsitution tab, helping finding the best time for subsitution change  #294
  - [FIX] all match orders are now consider in minute-bases prediction rating (behaviour change, position wap ...). Until now only substitution were considered #385
  - [FIX] counterattackrating calculation beyond divine has been fixed #398
-
 
 
 ### Matches
@@ -77,7 +79,6 @@ Changelist HO! 3.0
 
    - [NEW] information about new league will be available as soon as last game of the season has been played #247
    - [FIX] HO! will now automatically recognize league change at the begining of the season #391
-
 
 
 ## Translations
